Module qcgpu::gates

Gates and Gate Generation

Matrices are in row major format, and all gates use the num_complex::Complex<f32>; datatype.

Structs

Gate
Representation of a gate

Functions

h
Hadamard Gate

id
Identity gate

negh
Negative Hadamard Gate

r
The Phase Shift / R gate. Note that this function has an argument of angle, which corresponds to the rotation angle of the gate.

s
S / Phase Gate

t
T Gate

x
Pauli X / NOT Gate

y
Pauli Y Gate

z
Pauli Z Gate